% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliver.R
\name{deliver}
\alias{deliver}
\alias{deliver.default}
\alias{deliver.character}
\alias{deliver.www_theguardian_com}
\alias{deliver.www_huffingtonpost_co_uk}
\alias{deliver.www_buzzfeed_com}
\alias{deliver.www_forbes_com}
\title{Deliver online news articles}
\usage{
deliver(url, ...)

\method{deliver}{default}(url, ...)

\method{deliver}{character}(url, ...)

\method{deliver}{www_theguardian_com}(url, ...)

\method{deliver}{www_huffingtonpost_co_uk}(url, ...)

\method{deliver}{www_buzzfeed_com}(url, ...)

\method{deliver}{www_forbes_com}(url, ...)
}
\arguments{
\item{url}{The URL of the web article.}

\item{...}{Passed on to respective scraper.}
}
\value{
A data.frame (tibble) with media data and full text.
}
\description{
This function will determine the website of the urls given to it
  and call the appropriate webscraper.
}
