% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandurls.R
\name{expandurls}
\alias{expandurls}
\title{Expand URLs}
\usage{
expandurls(url, timeout = 15, ignore_fails = FALSE, ...)
}
\arguments{
\item{url}{Character object with URLs.}

\item{timeout}{How long should the function wait for the connection (in
seconds). If the query finishes earlier, results are returned immediately.}

\item{ignore_fails}{normally the function errors when a url can't be reached
due to connection issues. Setting to TRUE ignores this.}

\item{...}{Currently not used}
}
\value{
Character object with full (i.e., unshortened) URLs.
}
\description{
Expand URLs
}
