% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb_available}
\alias{pb_available}
\title{Show available scrapers}
\usage{
pb_available()
}
\value{
A character vector of supported domains.
}
\description{
Show available scrapers
}
\examples{
pb_available()
}
