% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliver.R
\name{pb_deliver_paper}
\alias{pb_deliver_paper}
\title{internal function to deliver specific newspapers}
\usage{
pb_deliver_paper(x, verbose, pb, ...)
}
\arguments{
\item{x}{Either a vector of URLs or a data.frame returned by
\link{pb_collect} or \link{pb_supply}.}

\item{verbose}{\code{FALSE} turns deliver silent. \code{TRUE} prints status
messages and a progress bar on the screen. \code{2L} turns on debug mode.
If \code{NULL} will be determined from
\code{getOption("paperboy_verbose")}.}

\item{pb}{a progress bar object.}

\item{...}{Passed on to \link{pb_collect}.}
}
\description{
internal function to deliver specific newspapers
}
\keyword{internal}
