% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliver_usatoday_com.R
\name{pb_deliver_paper.usatoday_com}
\alias{pb_deliver_paper.usatoday_com}
\title{internal function to deliver specific newspapers}
\usage{
\method{pb_deliver_paper}{usatoday_com}(x, verbose = NULL, ...)
}
\arguments{
\item{x}{A data.frame returned by  \link{pb_collect} with an additional class
indicating the domain of all links.}

\item{verbose}{A logical flag indicating whether information should be
printed to the screen. If \code{NULL} will be determined from
\code{getOption("paperboy_verbose")}.}

\item{...}{Passed on to \link{pb_collect}.}
}
\description{
internal function to deliver specific newspapers
}
\keyword{internal}
