% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{pb_inspect}
\alias{pb_inspect}
\title{Inspect content collected with pb_collect}
\usage{
pb_inspect(x, i = 1L, host_ip = "127.0.0.1", port = httpuv::randomPort(), ...)
}
\arguments{
\item{x}{a data.frame returned by \link{pb_collect}.}

\item{i}{which entry to display.}

\item{host_ip, port}{host IP and port to create the temporary web server that
shows the content.}

\item{...}{not used.}
}
\description{
Opens a browser to display the content saved in a row of a data.frame created
with \link{pb_collect}.
}
