% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{pb_new}
\alias{pb_new}
\title{Create new scraper}
\usage{
pb_new(np, author = "", issue = "")
}
\arguments{
\item{np}{domain or a URL of the newspaper this scraper is for.}

\item{author}{who wrote it.}

\item{issue}{is there a GitHub issue?}
}
\description{
Create new scraper
}
\examples{
\dontrun{
paperboy:::pb_new(np = "https://www.buzzfeed.com/",
                  author = "[@JBGruber](https://github.com/JBGruber/)")

paperboy:::pb_new_done()
}
}
\keyword{internal}
