% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_all}
\alias{replace_all}
\title{base R version of stringi::stri_replace_all()}
\usage{
replace_all(str, pattern, replacement, fixed = TRUE)
}
\description{
base R version of stringi::stri_replace_all()
}
\keyword{internal}
