% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dev.R
\name{test_parser}
\alias{test_parser}
\title{Test a Parser}
\usage{
test_parser(test_data)
}
\arguments{
\item{test_data}{A data frame of raw content.}
}
\value{
A success or failure message.
}
\description{
Test a parser using a data frame from \link{pb_collect}.
}
