% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliver.R, R/deliver.huffingtonpost.co.uk.R,
%   R/deliver.huffingtonpost.com.R, R/deliver.theguardian.com.R
\name{deliver}
\alias{deliver}
\alias{deliver.default}
\alias{deliver.character}
\alias{deliver.www_buzzfeed_com}
\alias{deliver.www_forbes_com}
\alias{deliver.www_huffingtonpost_co_uk}
\alias{deliver.www_huffpost_com}
\alias{deliver.www_theguardian_com}
\title{Deliver online news articles}
\usage{
deliver(url, verbose = TRUE, ...)

\method{deliver}{default}(url, verbose = TRUE, ...)

\method{deliver}{character}(url, verbose = TRUE, ...)

\method{deliver}{www_buzzfeed_com}(url, verbose = TRUE, ...)

\method{deliver}{www_forbes_com}(url, verbose = TRUE, ...)

\method{deliver}{www_huffingtonpost_co_uk}(url, verbose = TRUE, ...)

\method{deliver}{www_huffpost_com}(url, verbose = TRUE, ...)

\method{deliver}{www_huffingtonpost_co_uk}(url, verbose = TRUE, ...)

\method{deliver}{www_theguardian_com}(url, verbose = TRUE, ...)
}
\arguments{
\item{url}{The URL of the web article.}

\item{verbose}{A logical flag indicating whether information should be
printed to the screen.}

\item{...}{Passed on to respective scraper.}
}
\value{
A data.frame (tibble) with media data and full text.
}
\description{
This function will determine the website of the urls given to it
  and call the appropriate webscraper.
}
