% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_search.R
\name{html_search}
\alias{html_search}
\title{Search raw html for attributes}
\usage{
html_search(html, selectors, attributes = NULL, n = 1L)
}
\arguments{
\item{html}{raw html}

\item{selectors}{a vector of CSS selectors to include in search.}

\item{attributes}{attributes to extract. If NULL, returns text.}

\item{n}{if multiple are found, how many to return}
}
\value{
a vector of max length n
}
\description{
Search raw html for attributes
}
\keyword{internal}
