% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.r
\name{pb_collect_rss}
\alias{pb_collect_rss}
\title{Collect RSS feed}
\usage{
pb_collect_rss(x, ...)
}
\arguments{
\item{x}{URL(s) to RSS or Atom feed(s).}

\item{...}{passed to pb_collect.}
}
\value{
a character vector of URLs to articles
}
\description{
Collect the URLs of articles from RSS or Atom feed(s)
}
\examples{
pb_collect_rss("https://feeds.washingtonpost.com/rss/world")
}
