% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliver.R
\name{pb_deliver}
\alias{pb_deliver}
\title{Deliver online news articles}
\usage{
pb_deliver(x, try_default = TRUE, verbose = NULL, ...)
}
\arguments{
\item{x}{Either a vector of URLs or a data.frame returned by
\link{pb_collect}.}

\item{try_default}{if no parser is available, should a generic parser be used
\code{TRUE} or should the URL be skipped \code{FALSE}?}

\item{verbose}{\code{FALSE} turns deliver silent. \code{TRUE} prints status
messages and a progress bar on the screen. \code{2L} turns on debug mode.
If \code{NULL} will be determined from
\code{getOption("paperboy_verbose")}.}

\item{...}{Passed on to \link{pb_collect}.}
}
\value{
A data.frame (tibble) with media data and full text.
}
\description{
This function will determine the website of the urls given to it
  and call the appropriate webscraper.
}
