% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rss.r
\name{pb_find_rss}
\alias{pb_find_rss}
\title{Find RSS feed on a newspapers website}
\usage{
pb_find_rss(x)
}
\arguments{
\item{x}{main domain of the newspaper site to check for RSS feeds.}
}
\value{
A URL to the RSS feed(s) or NULL if nothing is found
}
\description{
Find RSS feed on a newspapers website
}
\examples{
pb_find_rss("https://www.buzzfeed.com/")
}
