% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pb_new}
\alias{pb_new}
\title{Create new scraper}
\usage{
pb_new(np)
}
\arguments{
\item{np}{domain of the newspaper this scraper is for.}
}
\description{
Create new scraper
}
\examples{
\dontrun{paperboy:::pb_new(np = "https://www.buzzfeed.com/")}
}
\keyword{internal}
