% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cookies.R
\name{pb_read_cookies}
\alias{pb_read_cookies}
\title{Read in cookie file}
\usage{
pb_read_cookies(...)
}
\arguments{
\item{...}{not used.}
}
\description{
Deprecated in favour of \link[cookiemonster]{add_cookies}.
}
