% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{pb_supply}
\alias{pb_supply}
\title{Supply paperboy with existing crawl data}
\usage{
pb_supply(
  urls,
  contents_raw,
  expanded_urls = NULL,
  statuses = rep(200L, length(urls)),
  collected_at = lubridate::now()
)
}
\arguments{
\item{urls}{Character vector with original URLs.}

\item{contents_raw}{Character vector with raw HTML fetched from these URLs.}

\item{expanded_urls}{Character vector with final (normalized and/or
redirected) URLs. If \code{NULL}, \link[curl]{curl_parse_url} is used to
normalize the URLs in \code{urls}.}

\item{statuses}{Integer vector with the HTTP response codes. The default
assumption is 200 ("OK") for for all pages.}

\item{collected_at}{The \code{POSIXct} time when the data were collected.
Note that paperboy records a single time per data.frame, not per URL.}
}
\value{
A data.frame (tibble) equivalent to that returned from
  \link{pb_collect}.
}
\description{
This function lets you use \link{pb_deliver} to parse already-collected raw
HTML by constructing a suitable data.frame. It is broadly equivalent to
\link{pb_collect}, but collects no data.
}
\examples{
pb_supply(
  urls = c('bbc.co.uk/news/articles/cz7dlgejqylo',
           'bbc.co.uk/news/articles/c1l1r1zne1ro'),
  contents_raw = c('<!DOCTYPE html><html lang="en-GB" class="no-js"><he...',
                   '<!DOCTYPE html><html lang="en-GB"><head><meta name=...'),
  expanded_urls = c('https://www.bbc.co.uk/news/articles/cz7dlgejqylo',
                    'https://www.bbc.co.uk/news/articles/c1l1r1zne1ro'),
  collected_at = lubridate::ymd_hms('2026-08-20 11:14:05')
)

}
\seealso{
\link{pb_collect}
}
