% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supply.R
\name{pb_supply}
\alias{pb_supply}
\title{Supply paperboy with existing crawl data}
\usage{
pb_supply(
  data,
  url = "url",
  content_raw = NULL,
  content_filenames = NULL,
  expanded_url = NULL,
  domain = NULL,
  status = NULL,
  collected_at = lubridate::now()
)
}
\arguments{
\item{data}{A data.frame with input data. At a minimum this must contain
a column of original urls and one of raw HTML data fetched from these URLs.}

\item{url}{The column name containing the original URLs.}

\item{content_raw}{A column name containing the raw HTML fetched from
these URLs. Exactly one of \code{content_raw} and \code{content_filenames}
must be provided.}

\item{content_filenames}{A column name containing filenames for the raw HTML
data fetched from these URLs. Exactly one of \code{content_raw} and
\code{content_filenames} must be provided.}

\item{expanded_url}{The column name containing final (normalized and/or
redirected) URLs. If \code{NULL}, \link[curl]{curl_parse_url} is used to
normalize the URLs in \code{data[[urls]]}.}

\item{domain}{The column name containing the domain to be used to select a
parser by \link{pb_deliver}. If \code{NULL}, the correct domains are
calculated.}

\item{status}{The column name containing HTTP response codes. If \code{NULL},
200 ("OK") is assumed for for all pages.}

\item{collected_at}{The \code{POSIXct} time when the data were collected.
Note that paperboy records a single time per data.frame, not per URL.}
}
\value{
A data.frame (tibble) equivalent to that returned from
  \link{pb_collect}.
}
\description{
This function lets you use \link{pb_deliver} to parse already-collected raw
HTML by constructing a suitable data.frame. It is broadly equivalent to
\link{pb_collect}, but collects no data.
}
\details{
If \code{content_raw} is provided, content will be kept in memory, as with
\code{pb_collect(save_dir = NULL)}. If \code{content_filenames} is provided
then the output data.frame will contain links to the data files, as with
\code{pb_collect(save_dir = 'some_directory')}.
}
\examples{
df <- dplyr::tibble(
  url = c('bbc.co.uk/news/articles/cz7dlgejqylo',
          'bbc.co.uk/news/articles/c1l1r1zne1ro'),
  html_data = c('<!DOCTYPE html><html lang="en-GB" class="no-js"><he...',
                '<!DOCTYPE html><html lang="en-GB"><head><meta name=...'),
  expanded_url = c('https://www.bbc.co.uk/news/articles/cz7dlgejqylo',
                   'https://www.bbc.co.uk/news/articles/c1l1r1zne1ro')
)
pb_supply(df,
          content_raw = 'html_data',
          expanded_url = 'expanded_url',
          collected_at = lubridate::ymd_hms('2026-08-20 11:14:05')
)

}
